refactor(themes): split color-themes.ts into per-theme modules
color-themes.ts had grown to 613 lines, ~430 of which were theme data.
Adding a theme meant editing the union type, three palette consts, and the
registry, all in one file.

The implementation now lives in src/themes/color-theme/: one module per
theme under definitions/, the shared neutral palette under palettes/, the
type contract in types.ts, and the apply/persist machinery in runtime.ts.
color-themes.ts becomes a barrel, so all nine importers are untouched.

Adding a theme is now: a new file in definitions/, one line in the
ColorThemeKey union, one line in definitions/index.ts.

Pure code motion — no behavior change. Every color value was moved by line
slice rather than retyped, and the result was verified by importing the
previous COLOR_THEMES alongside the new one and asserting deep equality
plus identical JSON.stringify output (which also catches key-order drift).

Note: solarized-light imports LIGHT_PLUS_TOKENS. Solarized Light only
overrides the tokens that differ from Light+, a dependency that was an
implicit spread in the monolith and is now an explicit import.
