+++ src/themes/color-theme/definitions/light-plus.ts
@@ -1,0 +1,130 @@
+import type { AgentServerUICssVariableName } from "#/styles/agent-server-ui-style-scope";
+import type { ColorThemeDefinition } from "../types";
+
+const LIGHT_PLUS_SCALE = {
+  "--cool-grey-50": "#181818",
+  "--cool-grey-100": "#1F1F1F",
+  "--cool-grey-200": "#424242",
+  "--cool-grey-300": "#4D4D4D",
+  "--cool-grey-400": "#616161",
+  "--cool-grey-500": "#767676",
+  "--cool-grey-600": "#8A8A8A",
+  "--cool-grey-700": "#D4D4D4",
+  "--cool-grey-800": "#E5E5E5",
+  "--cool-grey-900": "#EDEDED",
+  "--cool-grey-925": "#F3F3F3",
+  "--cool-grey-950": "#FFFFFF",
+  "--cool-grey-975": "#F8F8F8",
+};
+
+const LIGHT_PLUS_HEROUI = {
+  "--heroui-background": "0 0% 100%",
+  "--heroui-background-foreground": "0 0% 12.16%",
+  "--heroui-foreground-50": "0 0% 98%",
+  "--heroui-foreground-100": "0 0% 95.29%",
+  "--heroui-foreground-200": "0 0% 90.98%",
+  "--heroui-foreground-300": "0 0% 83.14%",
+  "--heroui-foreground-400": "0 0% 64%",
+  "--heroui-foreground-500": "0 0% 46.27%",
+  "--heroui-foreground-600": "0 0% 38.04%",
+  "--heroui-foreground-700": "0 0% 25.88%",
+  "--heroui-foreground-800": "0 0% 18%",
+  "--heroui-foreground-900": "0 0% 12.16%",
+  "--heroui-foreground": "0 0% 12.16%",
+  "--heroui-content1": "0 0% 100%",
+  "--heroui-content1-foreground": "0 0% 12.16%",
+  "--heroui-content2": "0 0% 97%",
+  "--heroui-content2-foreground": "0 0% 12.16%",
+  "--heroui-content3": "0 0% 95.29%",
+  "--heroui-content3-foreground": "0 0% 12.16%",
+  "--heroui-content4": "0 0% 90.98%",
+  "--heroui-content4-foreground": "0 0% 12.16%",
+  "--heroui-default-50": "0 0% 98%",
+  "--heroui-default-100": "0 0% 95.29%",
+  "--heroui-default-200": "0 0% 90.98%",
+  "--heroui-default-300": "0 0% 83.14%",
+  "--heroui-default-400": "0 0% 64%",
+  "--heroui-default-500": "0 0% 46.27%",
+  "--heroui-default-600": "0 0% 38.04%",
+  "--heroui-default-700": "0 0% 25.88%",
+  "--heroui-default-800": "0 0% 18%",
+  "--heroui-default-900": "0 0% 12.16%",
+  "--heroui-default-foreground": "0 0% 12.16%",
+  "--heroui-default": "0 0% 90.98%",
+};
+
+export const LIGHT_PLUS_TOKENS: Partial<
+  Record<AgentServerUICssVariableName, string>
+> = {
+  "--oh-color-primary": "#007ACC",
+  "--oh-color-logo": "#6B5700",
+  "--oh-color-base": "#FFFFFF",
+  "--oh-color-base-secondary": "#F3F3F3",
+  "--oh-color-danger": "#D13438",
+  "--oh-color-success": "#16825D",
+  "--oh-color-basic": "#616161",
+  "--oh-color-tertiary": "#E5E5E5",
+  "--oh-color-tertiary-light": "#4D4D4D",
+  "--oh-color-content": "#242424",
+  "--oh-color-content-2": "#181818",
+  "--oh-background": "#FFFFFF",
+  "--oh-foreground": "#1F1F1F",
+  "--oh-surface": "#F3F3F3",
+  "--oh-surface-foreground": "#1F1F1F",
+  "--oh-surface-raised": "#FFFFFF",
+  "--oh-surface-deep": "#E8E8E8",
+  "--oh-overlay": "#FFFFFF",
+  "--oh-overlay-foreground": "#1F1F1F",
+  "--oh-modal-title-foreground": "#1F1F1F",
+  "--oh-muted": "#616161",
+  "--oh-text-secondary": "#424242",
+  "--oh-text-tertiary": "#4D4D4D",
+  "--oh-text-dim": "#767676",
+  "--oh-text-subtle": "#616161",
+  "--oh-interactive-hover": "#E5E5E5",
+  "--oh-interactive-hover-low": "#F2F2F2",
+  "--oh-interactive-active": "#E8E8E8",
+  "--oh-interactive-selected": "#D6E9F8",
+  "--oh-context-window-foreground": "#1F1F1F",
+  "--oh-context-window-track-weight": "52%",
+  "--oh-scrollbar": "rgba(31, 31, 31, 0.24)",
+  "--oh-scrollbar-hover": "rgba(31, 31, 31, 0.4)",
+  "--oh-default": "#E5E5E5",
+  "--oh-default-foreground": "#1F1F1F",
+  "--oh-accent": "#007ACC",
+  "--oh-accent-foreground": "#FFFFFF",
+  "--oh-success": "#16825D",
+  "--oh-success-foreground": "#FFFFFF",
+  "--oh-warning": "#BF8803",
+  "--oh-warning-foreground": "#1F1F1F",
+  "--oh-danger": "#D13438",
+  "--oh-danger-foreground": "#FFFFFF",
+  "--oh-segment": "#F3F3F3",
+  "--oh-segment-foreground": "#1F1F1F",
+  "--oh-border": "#D4D4D4",
+  "--oh-border-input": "#C8C8C8",
+  "--oh-border-subtle": "#E5E5E5",
+  "--oh-separator": "rgba(31, 31, 31, 0.16)",
+  "--oh-focus": "#007ACC",
+  "--oh-status-success": "#16825D",
+  "--oh-status-error": "#D13438",
+  "--oh-link": "#006AB1",
+  "--oh-bg-dark": "#FFFFFF",
+  "--oh-bg-light": "#F3F3F3",
+  "--oh-bg-input": "#FFFFFF",
+  "--oh-bg-workspace": "#F8F8F8",
+  "--oh-text-editor-base": "#616161",
+  "--oh-text-editor-active": "#1F1F1F",
+  "--oh-bg-editor-sidebar": "#F3F3F3",
+  "--oh-bg-editor-active": "#E8E8E8",
+  "--oh-border-editor-sidebar": "#D4D4D4",
+  "--oh-bg-neutral-muted": "rgba(31, 31, 31, 0.08)",
+};
+
+export const lightPlus: ColorThemeDefinition = {
+  label: "Light+",
+  appearance: "light",
+  scale: LIGHT_PLUS_SCALE,
+  heroui: LIGHT_PLUS_HEROUI,
+  tokens: LIGHT_PLUS_TOKENS,
+};
