+++ src/themes/color-theme/definitions/openhands-deepsea.ts
@@ -1,0 +1,60 @@
+import type { ColorThemeDefinition } from "../types";
+
+export const openhandsDeepsea: ColorThemeDefinition = {
+  label: "OpenHands-DeepSea",
+  appearance: "dark",
+  // Matches the values already set by index.css; included so switching back
+  // from another theme restores the original palette explicitly.
+  scale: {
+    "--cool-grey-50": "#F7F9FC",
+    "--cool-grey-100": "#EEF2F7",
+    "--cool-grey-200": "#DCE3EE",
+    "--cool-grey-300": "#C3CDDC",
+    "--cool-grey-400": "#A3B0C4",
+    "--cool-grey-500": "#7E8A9E",
+    "--cool-grey-600": "#626D82",
+    "--cool-grey-700": "#4B5468",
+    "--cool-grey-800": "#383F50",
+    "--cool-grey-900": "#2C313F",
+    "--cool-grey-925": "#21252F",
+    "--cool-grey-950": "#0B0E14",
+    "--cool-grey-975": "#05070A",
+  },
+  // Values generated by heroui() from hero.ts — restore them explicitly when
+  // switching back from another theme.
+  heroui: {
+    "--heroui-background": "220 29.03% 6.08%",
+    "--heroui-background-foreground": "216 45.45% 97.84%",
+    "--heroui-foreground-50": "216 33.33% 2.94%",
+    "--heroui-foreground-100": "220 29.03% 6.08%",
+    "--heroui-foreground-200": "222.86 17.5% 15.69%",
+    "--heroui-foreground-300": "224.21 17.76% 20.98%",
+    "--heroui-foreground-400": "222.5 17.65% 26.67%",
+    "--heroui-foreground-500": "221.38 16.2% 35.1%",
+    "--heroui-foreground-600": "219.38 14.04% 44.71%",
+    "--heroui-foreground-700": "217.5 14.16% 55.69%",
+    "--heroui-foreground-800": "216.36 21.85% 70.39%",
+    "--heroui-foreground-900": "216 26.32% 81.37%",
+    "--heroui-foreground": "216 26.32% 81.37%",
+    "--heroui-content1": "222.86 17.5% 15.69%",
+    "--heroui-content1-foreground": "213.33 36% 95.1%",
+    "--heroui-content2": "224.21 17.76% 20.98%",
+    "--heroui-content2-foreground": "216.67 34.62% 89.8%",
+    "--heroui-content3": "222.5 17.65% 26.67%",
+    "--heroui-content3-foreground": "216 26.32% 81.37%",
+    "--heroui-content4": "221.38 16.2% 35.1%",
+    "--heroui-content4-foreground": "216.36 21.85% 70.39%",
+    "--heroui-default-50": "216 33.33% 2.94%",
+    "--heroui-default-100": "220 29.03% 6.08%",
+    "--heroui-default-200": "222.86 17.5% 15.69%",
+    "--heroui-default-300": "224.21 17.76% 20.98%",
+    "--heroui-default-400": "222.5 17.65% 26.67%",
+    "--heroui-default-500": "221.38 16.2% 35.1%",
+    "--heroui-default-600": "219.38 14.04% 44.71%",
+    "--heroui-default-700": "217.5 14.16% 55.69%",
+    "--heroui-default-800": "216.36 21.85% 70.39%",
+    "--heroui-default-900": "216 26.32% 81.37%",
+    "--heroui-default-foreground": "216 45.45% 97.84%",
+    "--heroui-default": "222.5 17.65% 26.67%",
+  },
+};
