+++ src/themes/color-theme/definitions/solarized-light.ts
@@ -1,0 +1,133 @@
+import type { AgentServerUICssVariableName } from "#/styles/agent-server-ui-style-scope";
+import type { ColorThemeDefinition } from "../types";
+// Solarized Light inherits Light+ token values, overriding only what differs.
+import { LIGHT_PLUS_TOKENS } from "./light-plus";
+
+const SOLARIZED_LIGHT_SCALE = {
+  "--cool-grey-50": "#002B36",
+  "--cool-grey-100": "#586E75",
+  "--cool-grey-200": "#657B83",
+  "--cool-grey-300": "#657B83",
+  "--cool-grey-400": "#839496",
+  "--cool-grey-500": "#93A1A1",
+  "--cool-grey-600": "#B7B1A1",
+  "--cool-grey-700": "#D7D0BC",
+  "--cool-grey-800": "#E6DFCC",
+  "--cool-grey-900": "#EEE8D5",
+  "--cool-grey-925": "#F7F0DC",
+  "--cool-grey-950": "#FDF6E3",
+  "--cool-grey-975": "#E9E2CF",
+};
+
+const SOLARIZED_LIGHT_HEROUI = {
+  "--heroui-background": "43.85 86.67% 94.12%",
+  "--heroui-background-foreground": "196 12.93% 45.49%",
+  "--heroui-foreground-50": "43.85 86.67% 94.12%",
+  "--heroui-foreground-100": "45.6 42.37% 88.43%",
+  "--heroui-foreground-200": "44 30% 82%",
+  "--heroui-foreground-300": "44 20% 75%",
+  "--heroui-foreground-400": "180 6.93% 60.39%",
+  "--heroui-foreground-500": "186.32 8.3% 55.1%",
+  "--heroui-foreground-600": "196 12.93% 45.49%",
+  "--heroui-foreground-700": "194.48 14.15% 40.2%",
+  "--heroui-foreground-800": "192.2 80.82% 14.31%",
+  "--heroui-foreground-900": "192.22 100% 10.59%",
+  "--heroui-foreground": "196 12.93% 45.49%",
+  "--heroui-content1": "43.85 86.67% 94.12%",
+  "--heroui-content1-foreground": "196 12.93% 45.49%",
+  "--heroui-content2": "45.6 42.37% 88.43%",
+  "--heroui-content2-foreground": "196 12.93% 45.49%",
+  "--heroui-content3": "44 30% 84%",
+  "--heroui-content3-foreground": "194.48 14.15% 40.2%",
+  "--heroui-content4": "44 24% 79%",
+  "--heroui-content4-foreground": "192.2 80.82% 14.31%",
+  "--heroui-default-50": "43.85 86.67% 94.12%",
+  "--heroui-default-100": "45.6 42.37% 88.43%",
+  "--heroui-default-200": "44 30% 82%",
+  "--heroui-default-300": "44 20% 75%",
+  "--heroui-default-400": "180 6.93% 60.39%",
+  "--heroui-default-500": "186.32 8.3% 55.1%",
+  "--heroui-default-600": "196 12.93% 45.49%",
+  "--heroui-default-700": "194.48 14.15% 40.2%",
+  "--heroui-default-800": "192.2 80.82% 14.31%",
+  "--heroui-default-900": "192.22 100% 10.59%",
+  "--heroui-default-foreground": "192.22 100% 10.59%",
+  "--heroui-default": "45.6 42.37% 88.43%",
+};
+
+const SOLARIZED_LIGHT_TOKENS: Partial<
+  Record<AgentServerUICssVariableName, string>
+> = {
+  ...LIGHT_PLUS_TOKENS,
+  "--oh-color-primary": "#268BD2",
+  "--oh-color-logo": "#B58900",
+  "--oh-color-base": "#FDF6E3",
+  "--oh-color-base-secondary": "#EEE8D5",
+  "--oh-color-danger": "#DC322F",
+  "--oh-color-success": "#859900",
+  "--oh-color-basic": "#839496",
+  "--oh-color-tertiary": "#E6DFCC",
+  "--oh-color-tertiary-light": "#657B83",
+  "--oh-color-content": "#657B83",
+  "--oh-color-content-2": "#002B36",
+  "--oh-background": "#FDF6E3",
+  "--oh-foreground": "#657B83",
+  "--oh-surface": "#EEE8D5",
+  "--oh-surface-foreground": "#586E75",
+  "--oh-surface-raised": "#FDF6E3",
+  "--oh-surface-deep": "#E6DFCC",
+  "--oh-overlay": "#FDF6E3",
+  "--oh-overlay-foreground": "#586E75",
+  "--oh-modal-title-foreground": "#002B36",
+  "--oh-muted": "#839496",
+  "--oh-text-secondary": "#657B83",
+  "--oh-text-tertiary": "#586E75",
+  "--oh-text-dim": "#839496",
+  "--oh-text-subtle": "#657B83",
+  "--oh-interactive-hover": "#E6DFCC",
+  "--oh-interactive-hover-low": "#F5EEDB",
+  "--oh-interactive-active": "#DDD6C3",
+  "--oh-interactive-selected": "#D7E8EA",
+  "--oh-context-window-foreground": "#002B36",
+  "--oh-context-window-track-weight": "54%",
+  "--oh-scrollbar": "rgba(88, 110, 117, 0.28)",
+  "--oh-scrollbar-hover": "rgba(88, 110, 117, 0.45)",
+  "--oh-default": "#E6DFCC",
+  "--oh-default-foreground": "#586E75",
+  "--oh-accent": "#268BD2",
+  "--oh-accent-foreground": "#FFFFFF",
+  "--oh-success": "#859900",
+  "--oh-success-foreground": "#002B36",
+  "--oh-warning": "#B58900",
+  "--oh-warning-foreground": "#002B36",
+  "--oh-danger": "#DC322F",
+  "--oh-danger-foreground": "#FFFFFF",
+  "--oh-segment": "#EEE8D5",
+  "--oh-segment-foreground": "#586E75",
+  "--oh-border": "#D7D0BC",
+  "--oh-border-input": "#C8C1AE",
+  "--oh-border-subtle": "#E6DFCC",
+  "--oh-separator": "rgba(88, 110, 117, 0.22)",
+  "--oh-focus": "#268BD2",
+  "--oh-status-success": "#859900",
+  "--oh-status-error": "#DC322F",
+  "--oh-link": "#268BD2",
+  "--oh-bg-dark": "#FDF6E3",
+  "--oh-bg-light": "#EEE8D5",
+  "--oh-bg-input": "#FDF6E3",
+  "--oh-bg-workspace": "#F7F0DC",
+  "--oh-text-editor-base": "#839496",
+  "--oh-text-editor-active": "#586E75",
+  "--oh-bg-editor-sidebar": "#EEE8D5",
+  "--oh-bg-editor-active": "#E6DFCC",
+  "--oh-border-editor-sidebar": "#D7D0BC",
+  "--oh-bg-neutral-muted": "rgba(88, 110, 117, 0.1)",
+};
+
+export const solarizedLight: ColorThemeDefinition = {
+  label: "Solarized Light",
+  appearance: "light",
+  scale: SOLARIZED_LIGHT_SCALE,
+  heroui: SOLARIZED_LIGHT_HEROUI,
+  tokens: SOLARIZED_LIGHT_TOKENS,
+};
