+++ src/themes/color-theme/palettes/neutral.ts
@@ -1,0 +1,71 @@
+// Shared neutral palette used by the OpenHands-Neutral and -Neo themes.
+
+// HSL channel strings for the neutral grey palette (H=0, S=0%, L=hex/255*100)
+// prettier-ignore
+const NEUTRAL_HSL = {
+  50:  "0 0% 96.86%", // #F7F7F7
+  100: "0 0% 92.55%", // #ECECEC
+  200: "0 0% 86.27%", // #DCDCDC
+  300: "0 0% 74.51%", // #BEBEBE
+  400: "0 0% 59.22%", // #979797
+  500: "0 0% 45.1%",  // #737373
+  600: "0 0% 33.73%", // #565656
+  700: "0 0% 25.1%",  // #404040
+  800: "0 0% 19.22%", // #313131
+  850: "0 0% 15.69%", // #282828
+  900: "0 0% 12.55%", // #202020
+  950: "0 0% 9.41%",  // #181818
+  975: "0 0% 6.27%",  // #101010
+};
+
+export const NEUTRAL_SCALE = {
+  "--cool-grey-50": "#F7F7F7",
+  "--cool-grey-100": "#ECECEC",
+  "--cool-grey-200": "#DCDCDC",
+  "--cool-grey-300": "#BEBEBE",
+  "--cool-grey-400": "#979797",
+  "--cool-grey-500": "#737373",
+  "--cool-grey-600": "#565656",
+  "--cool-grey-700": "#404040",
+  "--cool-grey-800": "#313131",
+  "--cool-grey-900": "#282828",
+  "--cool-grey-925": "#202020",
+  "--cool-grey-950": "#181818",
+  "--cool-grey-975": "#101010",
+};
+
+export const NEUTRAL_HEROUI = {
+  "--heroui-background": NEUTRAL_HSL[950],
+  "--heroui-background-foreground": NEUTRAL_HSL[50],
+  "--heroui-foreground-50": NEUTRAL_HSL[975],
+  "--heroui-foreground-100": NEUTRAL_HSL[950],
+  "--heroui-foreground-200": NEUTRAL_HSL[900],
+  "--heroui-foreground-300": NEUTRAL_HSL[850],
+  "--heroui-foreground-400": NEUTRAL_HSL[800],
+  "--heroui-foreground-500": NEUTRAL_HSL[700],
+  "--heroui-foreground-600": NEUTRAL_HSL[600],
+  "--heroui-foreground-700": NEUTRAL_HSL[500],
+  "--heroui-foreground-800": NEUTRAL_HSL[400],
+  "--heroui-foreground-900": NEUTRAL_HSL[300],
+  "--heroui-foreground": NEUTRAL_HSL[300],
+  "--heroui-content1": NEUTRAL_HSL[900],
+  "--heroui-content1-foreground": NEUTRAL_HSL[100],
+  "--heroui-content2": NEUTRAL_HSL[850],
+  "--heroui-content2-foreground": NEUTRAL_HSL[200],
+  "--heroui-content3": NEUTRAL_HSL[800],
+  "--heroui-content3-foreground": NEUTRAL_HSL[300],
+  "--heroui-content4": NEUTRAL_HSL[700],
+  "--heroui-content4-foreground": NEUTRAL_HSL[400],
+  "--heroui-default-50": NEUTRAL_HSL[975],
+  "--heroui-default-100": NEUTRAL_HSL[950],
+  "--heroui-default-200": NEUTRAL_HSL[900],
+  "--heroui-default-300": NEUTRAL_HSL[850],
+  "--heroui-default-400": NEUTRAL_HSL[800],
+  "--heroui-default-500": NEUTRAL_HSL[700],
+  "--heroui-default-600": NEUTRAL_HSL[600],
+  "--heroui-default-700": NEUTRAL_HSL[500],
+  "--heroui-default-800": NEUTRAL_HSL[400],
+  "--heroui-default-900": NEUTRAL_HSL[300],
+  "--heroui-default-foreground": NEUTRAL_HSL[50],
+  "--heroui-default": NEUTRAL_HSL[800],
+};
